+++ framework/src/play-java/src/main/scala/play/core/ObjectMapperModule.scala
@@ -6,8 +6,8 @@ package play.core
 import com.fasterxml.jackson.databind.ObjectMapper
 import play.api._
 import play.api.inject._
+import play.api.libs.json.jackson.PlayJsonModule
 import play.libs.Json
-import play.api.libs.json.JacksonJson
 
 import javax.inject._
 
@@ -29,7 +29,7 @@ class ObjectMapperModule extends Module {
 @Singleton
 class ObjectMapperProvider @Inject() (lifecycle: ApplicationLifecycle) extends Provider[ObjectMapper] {
   lazy val get = {
-    val objectMapper = JacksonJson.createMapper()
+    val objectMapper = new ObjectMapper().registerModule(PlayJsonModule)
     Json.setObjectMapper(objectMapper)
     lifecycle.addStopHook { () =>
       Future.successful(Json.setObjectMapper(null))
